requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
CacheControl==0.12.6
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0
distlib==0.3.1
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.2
django-adminlte-3==0.1.6
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.6.4
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
pipenv==2020.11.15
progress==1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pyparsing==2.4.6
python-decouple==3.4
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
virtualenv==20.2.2
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
